Flytt logikk til service
