+++ src/main/kotlin/no/nav/tilleggsstonader/sak/journalføring/JournalpostService.kt
@@ -12,11 +12,14 @@ import no.nav.tilleggsstonader.kontrakter.sak.DokumentBrevkode
 import no.nav.tilleggsstonader.kontrakter.søknad.Søknadsskjema
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.SøknadsskjemaBarnetilsyn
 import no.nav.tilleggsstonader.sak.fagsak.domain.Fagsak
+import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvisIkke
+import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
+import no.nav.tilleggsstonader.sak.opplysninger.pdl.PersonService
 import no.nav.tilleggsstonader.sak.vedlegg.VedleggRequest
 import org.springframework.stereotype.Service
 
 @Service
-class JournalpostService(private val journalpostClient: JournalpostClient) {
+class JournalpostService(private val journalpostClient: JournalpostClient, private val personService: PersonService) {
     fun finnJournalposterForBruker(personIdent: String, vedleggRequest: VedleggRequest): List<Journalpost> {
         return journalpostClient.finnJournalposterForBruker(
             JournalposterForBrukerRequest(
@@ -33,7 +36,10 @@ class JournalpostService(private val journalpostClient: JournalpostClient) {
         return journalpostClient.hentJournalpost(journalpostId)
     }
 
-    fun opprettJournalpost(arkiverDokumentRequest: ArkiverDokumentRequest, saksbehandler: String? = null): ArkiverDokumentResponse {
+    fun opprettJournalpost(
+        arkiverDokumentRequest: ArkiverDokumentRequest,
+        saksbehandler: String? = null,
+    ): ArkiverDokumentResponse {
         return journalpostClient.opprettJournalpost(arkiverDokumentRequest, saksbehandler)
     }
 
@@ -71,7 +77,11 @@ class JournalpostService(private val journalpostClient: JournalpostClient) {
         }
     }
 
-    private fun ferdigstillJournalføring(journalpostId: String, journalførendeEnhet: String, saksbehandler: String? = null) {
+    private fun ferdigstillJournalføring(
+        journalpostId: String,
+        journalførendeEnhet: String,
+        saksbehandler: String? = null,
+    ) {
         journalpostClient.ferdigstillJournalpost(journalpostId, journalførendeEnhet, saksbehandler)
     }
 
@@ -81,7 +91,8 @@ class JournalpostService(private val journalpostClient: JournalpostClient) {
         eksternFagsakId: Long,
         saksbehandler: String? = null,
     ) {
-        val oppdatertJournalpost = JournalføringHelper.lagOppdaterJournalpostRequest(journalpost, eksternFagsakId, dokumenttitler)
+        val oppdatertJournalpost =
+            JournalføringHelper.lagOppdaterJournalpostRequest(journalpost, eksternFagsakId, dokumenttitler)
         journalpostClient.oppdaterJournalpost(oppdatertJournalpost, journalpost.journalpostId, saksbehandler)
     }
 
@@ -90,11 +101,40 @@ class JournalpostService(private val journalpostClient: JournalpostClient) {
         return journalpostClient.hentSøknadTilsynBarn(søknadJournalpost.journalpostId, dokumentinfo.dokumentInfoId)
     }
 
+    fun finnJournalpostOgPersonIdent(journalpostId: String): Pair<Journalpost, String> {
+        val journalpost = hentJournalpost(journalpostId)
+        val personIdent = journalpost.bruker?.let {
+            when (it.type) {
+                BrukerIdType.FNR -> it.id
+                BrukerIdType.AKTOERID -> personService.hentPersonIdenter(it.id).gjeldende().ident
+                BrukerIdType.ORGNR -> error("Kan ikke hente journalpost=$journalpostId for orgnr")
+            }
+        } ?: error("Kan ikke hente journalpost=$journalpostId uten bruker")
+        return Pair(journalpost, personIdent)
+    }
+
     fun hentDokument(
-        journalpostId: String,
+        journalpost: Journalpost,
         dokumentInfoId: String,
         dokumentVariantformat: Dokumentvariantformat = Dokumentvariantformat.ARKIV,
     ): ByteArray {
-        return journalpostClient.hentDokument(journalpostId, dokumentInfoId, dokumentVariantformat)
+        validerDokumentKanHentes(journalpost, dokumentInfoId)
+        return journalpostClient.hentDokument(journalpost.journalpostId, dokumentInfoId, dokumentVariantformat)
+    }
+
+    private fun validerDokumentKanHentes(
+        journalpost: Journalpost,
+        dokumentInfoId: String,
+    ) {
+        val dokument = journalpost.dokumenter?.find { it.dokumentInfoId == dokumentInfoId }
+        feilHvis(dokument == null) {
+            "Finner ikke dokument med id=$dokumentInfoId for journalpost med id=${journalpost.journalpostId}"
+        }
+        brukerfeilHvisIkke(
+            dokument.dokumentvarianter?.any { it.variantformat == Dokumentvariantformat.ARKIV }
+                ?: false,
+        ) {
+            "Vedlegget er sannsynligvis under arbeid, må åpnes i gosys"
+        }
     }
 }
